refactor: revise Projects page
