+++ components/templates/Projects/Projects.tsx
@@ -11,8 +11,8 @@ import GoalSection from './GoalSection';
 import { Goals } from './GoalSection/goals';
 import CareerEvents from './CareerEvents/CareerEvents';
 import NextSection from 'components/atoms/NextSection';
-import BannerContainer, { BannerType } from 'components/molecules/BannerContainer';
 import TabNav from 'components/molecules/TabNav';
+import TemplateWrapper from 'components/organisms/TemplateWrapper';
 
 export type ProjectsType = typeof translateMap.projects;
 
@@ -37,14 +37,12 @@ const Projects: NextPage = () => {
   };
 
   return (
-    <>
-      <BannerContainer
-        type={BannerType.general}
-        primaryTitle="三大實作專案"
-        subTitle=""
-        description="資訊種子培訓計畫不同於大學一貫的授課方式，讓學員在實踐中學習。透過執行 4 大專案，參與 10+ 堂來自業界講師的課程，了解業界生態，並探索自己未來的職涯方向。
+    <TemplateWrapper
+      primaryTitle="三大實作專案"
+      subTitle=""
+      description="資訊種子培訓計畫不同於大學一貫的授課方式，讓學員在實踐中學習。透過執行 4 大專案，參與 10+ 堂來自業界講師的課程，了解業界生態，並探索自己未來的職涯方向。
 培養跨領域合作、解決問題的思維等職場必備的能力，成為能踏入職場的人才。"
-      />
+    >
       <div ref={elRef}>
         <TabNav tabs={tabs} activeTab={activeTab} onTabClick={changeTab} onStickyChange={setIsSubNavStuck} />
         {(() => {
@@ -94,7 +92,7 @@ const Projects: NextPage = () => {
           }
         })()}
       </div>
-    </>
+    </TemplateWrapper>
   );
 };
